loyalty_program/requirements.txt
djangorestframework==3.11.2
Django==3.1.8
django-cors-headers==3.5.0
pika==1.1.0